+++ extension/src-language-server/stpa/diagram/utils.ts
@@ -3,7 +3,7 @@
  *
  * http://rtsys.informatik.uni-kiel.de/kieler
  *
- * Copyright 2021 by
+ * Copyright 2021-2023 by
  * + Kiel University
  *   + Department of Computer Science
  *     + Real-Time and Embedded Systems Group
@@ -17,40 +17,22 @@
 
 import { AstNode } from "langium";
 import {
-    isHazard, isResponsibility, isSystemConstraint, isContConstraint, isSafetyConstraint, isUCA, isLossScenario,
-    isLoss, Hazard, SystemConstraint, isContext
+    Context,
+    Hazard,
+    Node,
+    SystemConstraint,
+    isContConstraint,
+    isContext,
+    isHazard,
+    isLossScenario,
+    isResponsibility,
+    isSafetyConstraint,
+    isSystemConstraint,
+    isUCA
 } from "../../generated/ast";
-import { groupValue } from "./synthesis-options";
 import { STPANode } from "./stpa-interfaces";
 import { STPAAspect } from "./stpa-model";
-
-/* export function determineLayerForCSNodes(nodes: CSNode[]): void {
-    let layer = nodes.length
-    let sinks: CSNode[] = []
-    while (true) {
-        for (let n of nodes) {
-            let s = true
-            for (let edge of n.outgoingEdges) {
-                if (edge instanceof CSEdge && edge.direction == EdgeDirection.DOWN && edge.target instanceof CSNode && !edge.target.layer) {
-                    s = false
-                }
-            }
-            if (s) {
-                sinks.push(n)
-                break;
-            }
-        }
-        if (sinks.length == 0) {
-            break;
-        }
-
-        for (let s of sinks) {
-            s.layer = layer
-        }
-        layer--
-        sinks = []
-    }
-} */
+import { groupValue } from "./synthesis-options";
 
 /**
  * Getter for the references contained in {@code node}.
@@ -84,37 +66,11 @@ export function getTargets(node: AstNode, hierarchy: boolean): AstNode[] {
 }
 
 /**
- * Getter for the aspect of a STPA component.
- * @param node AstNode which aspect should determined.
- * @returns the aspect of {@code node}.
- */
-export function getAspect(node: AstNode): STPAAspect {
-    if (isLoss(node)) {
-        return STPAAspect.LOSS;
-    } else if (isHazard(node)) {
-        return STPAAspect.HAZARD;
-    } else if (isSystemConstraint(node)) {
-        return STPAAspect.SYSTEMCONSTRAINT;
-    } else if (isUCA(node) || isContext(node)) {
-        return STPAAspect.UCA;
-    } else if (isResponsibility(node)) {
-        return STPAAspect.RESPONSIBILITY;
-    } else if (isContConstraint(node)) {
-        return STPAAspect.CONTROLLERCONSTRAINT;
-    } else if (isLossScenario(node)) {
-        return STPAAspect.SCENARIO;
-    } else if (isSafetyConstraint(node)) {
-        return STPAAspect.SAFETYREQUIREMENT;
-    }
-    return STPAAspect.UNDEFINED;
-}
-
-/**
  * Collects the {@code topElements}, their children, their children's children and so on.
  * @param topElements The top elements that possbible have children.
  * @returns A list with the given {@code topElements} and their descendants.
  */
-export function collectElementsWithSubComps(topElements: (Hazard | SystemConstraint)[]): AstNode[] {
+export function collectElementsWithSubComps(topElements: (Hazard | SystemConstraint)[]): (Hazard | SystemConstraint)[] {
     let result = topElements;
     let todo = topElements;
     for (let i = 0; i < todo.length; i++) {
@@ -126,7 +82,6 @@ export function collectElementsWithSubComps(topElements: (Hazard | SystemConstra
     }
     return result;
 }
-
 /**
  * Determines the layer {@code node} should be in depending on the STPA aspect it represents.
  * @param node STPANode for which the layer should be determined.
@@ -198,4 +153,89 @@ export function setLevelsForSTPANodes(nodes: STPANode[], groupUCAs: groupValue):
         const layer = determineLayerForSTPANode(node, maxHazardDepth, maxSysConsDepth, map, groupUCAs);
         node.level = -layer;
     }
+}
+
+
+/**
+ * Set the levels of the control structure nodes.
+ * @param nodes The nodes representing the control structure.
+ */
+export function setLevelOfCSNodes(nodes: Node[]): void {
+    const visited = new Map<string, Set<string>>();
+    for (const node of nodes) {
+        visited.set(node.name, new Set<string>());
+    }
+    nodes[0].level = 0;
+    assignLevel(nodes[0], visited);
+}
+
+/**
+ * Assigns the level to the connected nodes of {@code node}.
+ * @param node The node for which the connected nodes should be assigned a level.
+ * @param visited The edges that have been visited.
+ */
+function assignLevel(node: Node, visited: Map<string, Set<string>>): void {
+    for (const action of node.actions) {
+        const target = action.target.ref;
+        if (target && !visited.get(node.name)?.has(target.name)) {
+            visited.get(node.name)?.add(target.name);
+            if (target.level === undefined || target.level < node.level! + 1) {
+                target.level = node.level! + 1;
+            }
+            assignLevel(target, visited);
+        }
+    }
+    for (const feedback of node.feedbacks) {
+        const target = feedback.target.ref;
+        if (target && !visited.get(node.name)?.has(target.name)) {
+            visited.get(node.name)?.add(target.name);
+            if (target.level === undefined || target.level > node.level! - 1) {
+                target.level = node.level! - 1;
+            }
+            assignLevel(target, visited);
+        }
+    }
+}
+
+/**
+ * Creates a description for the given UCA context.
+ * @param uca The UCA context.
+ * @returns the description of the UCA context.
+ */
+export function createUCAContextDescription(uca: Context): string {
+    const rule = uca.$container;
+    const controlAction = rule.action.$refText;
+    let description = rule.system.$refText;
+    switch (rule.type) {
+        case 'not-provided':
+            description += " did not provide " + controlAction;
+            break;
+        case 'provided':
+            description += " provided " + controlAction;
+            break;
+        case 'too-late':
+            description += " provided " + controlAction + " too late";
+            break;
+        case 'too-early':
+            description += " provided " + controlAction + " too early";
+            break;
+        case 'wrong-time':
+            description += " provided " + controlAction + " at the wrong time";
+            break;
+        case 'applied-too-long':
+            description += " applied " + controlAction + " too long";
+            break;
+        case 'stopped-too-soon':
+            description += " stopped " + controlAction + " too soon";
+            break;
+    }
+    description += " in the context of ";
+    for (let i = 0; i < uca.vars.length; i++) {
+        description += uca.vars[i].$refText + "=" + uca.values[i];
+        if (i < uca.vars.length - 1) {
+            description += ", ";
+        }
+    }
+
+    return description;
 }
